Commented mobs.py, moved nodes to nodes.py
